requirements.txt
pyrogram
tgcrypto
pykeyboard
python-dotenv